+++ cmd/insprd/memory/tree/manager.go
@@ -85,36 +85,36 @@ func (mm *treeMemoryManager) GetTransactionChanges() (diff.Changelog, error) {
 	return cl, err
 }
 
-// RootGetter is a structure that gets components from the root, without the current changes.
-type RootGetter struct {
+// PermTreeGetter is a structure that gets components from the root, without the current changes.
+type PermTreeGetter struct {
 	tree *meta.App
 }
 
 // Apps returns a getter for apps on the root.
-func (t *RootGetter) Apps() AppGetInterface {
-	return &AppRootGetter{
+func (t *PermTreeGetter) Apps() AppGetInterface {
+	return &AppPermTreeGetter{
 		tree: t.tree,
 	}
 }
 
 // Channels returns a getter for channels on the root.
-func (t *RootGetter) Channels() ChannelGetInterface {
-	return &ChannelRootGetter{}
+func (t *PermTreeGetter) Channels() ChannelGetInterface {
+	return &ChannelPermTreeGetter{}
 }
 
 // Types returns a getter for Types on the root
-func (t *RootGetter) Types() TypeGetInterface {
-	return &TypeRootGetter{}
+func (t *PermTreeGetter) Types() TypeGetInterface {
+	return &TypePermTreeGetter{}
 }
 
 // Alias returns a getter for alias on the root
-func (t *RootGetter) Alias() AliasGetInterface {
-	return &AliasRootGetter{}
+func (t *PermTreeGetter) Alias() AliasGetInterface {
+	return &AliasPermTreeGetter{}
 }
 
-// Root returns a getter for objects on the root of the tree, without the current changes.
-func (mm *treeMemoryManager) Root() GetInterface {
-	return &RootGetter{
-		tree: mm.tree,
+// Tree returns a getter for objects on the tree without the current changes.
+func (tmm *treeMemoryManager) Tree() GetInterface {
+	return &PermTreeGetter{
+		tree: tmm.tree,
 	}
 }